Move events from the VM
