+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/list/ProductListViewModel.kt
@@ -21,12 +21,13 @@ import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.media.MediaFileUploadHandler
 import com.woocommerce.android.ui.products.ProductStatus
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.ScrollToTop
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.SelectProducts
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.ShowAddProductBottomSheet
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.ShowProductFilterScreen
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.ShowProductSortingBottomSheet
-import com.woocommerce.android.ui.products.list.ProductListViewModel.ProductListEvent.ShowUpdateDialog
+import com.woocommerce.android.ui.products.ai.banner.AIProductBannerDialogShouldBeShown
+import com.woocommerce.android.ui.products.list.ProductListEvent.ScrollToTop
+import com.woocommerce.android.ui.products.list.ProductListEvent.SelectProducts
+import com.woocommerce.android.ui.products.list.ProductListEvent.ShowAddProductBottomSheet
+import com.woocommerce.android.ui.products.list.ProductListEvent.ShowProductFilterScreen
+import com.woocommerce.android.ui.products.list.ProductListEvent.ShowProductSortingBottomSheet
+import com.woocommerce.android.ui.products.list.ProductListEvent.ShowUpdateDialog
 import com.woocommerce.android.util.IsWindowClassLargeThanCompact
 import com.woocommerce.android.util.WooLog
 import com.woocommerce.android.viewmodel.LiveDataDelegate
@@ -59,6 +60,7 @@ class ProductListViewModel @Inject constructor(
     private val selectedSite: SelectedSite,
     private val wooCommerceStore: WooCommerceStore,
     private val isWindowClassLargeThanCompact: IsWindowClassLargeThanCompact,
+    private val aiProductBannerDialogShouldBeShown: AIProductBannerDialogShouldBeShown
 ) : ScopedViewModel(savedState) {
     companion object {
         private const val KEY_PRODUCT_FILTER_OPTIONS = "key_product_filter_options"
@@ -109,6 +111,10 @@ class ProductListViewModel @Inject constructor(
         mediaFileUploadHandler.observeProductImageChanges()
             .onEach { loadProducts() }
             .launchIn(this)
+
+        if (aiProductBannerDialogShouldBeShown()) {
+            triggerEvent(MultiLiveEvent.Event.ShowAIProductBannerDialog)
+        }
     }
 
     override fun onCleared() {
@@ -763,40 +769,6 @@ class ProductListViewModel @Inject constructor(
 
         @IgnoredOnParcel
         val isFilteringActive = filterCount != null && filterCount > 0
-    }
-
-    sealed class ProductListEvent : MultiLiveEvent.Event() {
-        data object ScrollToTop : ProductListEvent()
-        data object ShowAddProductBottomSheet : ProductListEvent()
-        data object ShowProductSortingBottomSheet : ProductListEvent()
-        data class ShowProductFilterScreen(
-            val stockStatusFilter: String?,
-            val productTypeFilter: String?,
-            val productStatusFilter: String?,
-            val productCategoryFilter: String?,
-            val selectedCategoryName: String?
-        ) : ProductListEvent()
-        data class ShowProductUpdateStockStatusScreen(val productsIds: List<Long>) : ProductListEvent()
-        sealed class ShowUpdateDialog : ProductListEvent() {
-            abstract val productsIds: List<Long>
-
-            data class Price(override val productsIds: List<Long>) : ShowUpdateDialog()
-            data class Status(override val productsIds: List<Long>) : ShowUpdateDialog()
-        }
-        data class ShowDiscardProductChangesConfirmationDialog(
-            val productId: Long,
-            val productName: String,
-        ) : ProductListEvent()
-        data class OpenProduct(
-            val productId: Long,
-            val oldPosition: Int,
-            val newPosition: Int,
-            val sharedView: View?
-        ) : ProductListEvent()
-
-        data object OpenEmptyProduct : ProductListEvent()
-
-        data class SelectProducts(val productsIds: List<Long>) : ProductListEvent()
     }
 
     enum class ProductListState { Selecting, Browsing }
